modified:   .gitignore modified:   package-lock.json modified:   package.json modified:   src/components/Articles/ArticleEdit.jsx modified:   src/components/NavBar.jsx renamed:    src/components/DarkMode.jsx -> src/contexts/DarkMode.jsx renamed:    src/components/ProtectedRoute.jsx -> src/contexts/ProtectedRoute.jsx modified:   src/hooks/ConnApi.jsx modified:   src/routes/Home/Home.jsx new file:   src/routes/Home/Hooks/Conections.js modified:   src/routes/styles/HomeScreen.css
