+++ src/contexts/ProtectedRoute.jsx
@@ -1,8 +1,6 @@
 import { Navigate } from "react-router-dom";
 
-const ProtectedRoute = ({ children }) => {
+export const ProtectedRoute = ({ children }) => {
   const token = localStorage.getItem("authToken");
   return token ? children : <Navigate to="/login" />;
 };
-
-export default ProtectedRoute;
